requirements.txt
Flask
Pillow
requests
colorama